refactor: move some functions from driver to kernel_interface/common
